backend/CLIPParser/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.1.3
easyocr==1.7.0
requests==2.31.0
python-multipart==0.0.6
